Extract state history line chart data processing into a pure function
